+++ pennyway-client-iOS/pennyway-client-iOS/Extension/Date/DateExtensions.swift
@@ -1,0 +1,14 @@
+
+// MARK: - Date 익스텐션
+
+extension Date {
+    static let calendarDayDateFormatter: DateFormatter = {
+        let formatter = DateFormatter()
+        formatter.dateFormat = "MMMM yyyy dd"
+        return formatter
+    }()
+  
+    var formattedCalendarDayDate: String {
+        return Date.calendarDayDateFormatter.string(from: self)
+    }
+}
